Added location handling of database files
We've decided that configuration will be stored in the databases,
this means all we need for loading is the location of said
databases. We're storing this in a file named `dbloc` in a local
location according to spec.

The default database location is in a directory named `db` in the
same location as the `dbloc` file.

The location and name of the database directory are mutable, the
location of the `dbloc` is inmutable because we need to always be
able to find it on startup.