Added very basic (and naive) storage and retrieval of configuration
The current implementation uses encoding/json under the hood to
store whole values into the database. Callers should not assume
a specific encoding is used.

Naive implementation with just basic Load and Store methods, with
no real world usage of this API so far. There is no telling if this
kind of API is actually functional enough.