Removed ScheduleJobNow, Added Schedule
Motivation is that calling ScheduleJob with time.Now as first
argument is just as simple as ScheduleJobNow.

The new Schedule is for skipping the Task creation step in user
code. Since the current implementation of Task does not have any
extra setup available this method should be used in most cases.