Added initial result of data handling brainstorm
The initial idea is that the main process stores unique identifiers
to data. These identifiers contain enough information to both know
where to look for concrete data, and information as to what data we
are looking for.

Exposing data from a plugin to plethora is now as simple as saving
the unique identifier in the global database. The database should
keep track of some very small information for the data kept, such as
the time the data was added, if the data has been seen by the user,
at what time did the user see this data, and other possiblities.

The current interface should be enough to start preliminary work
on other parts of the system.